+++ homeassistant/util/yaml/__init__.py
@@ -1,0 +1,15 @@
+"""YAML utility functions."""
+from .const import (
+    SECRET_YAML, _SECRET_NAMESPACE
+)
+from .dumper import dump, save_yaml
+from .loader import (
+    clear_secret_cache, load_yaml, secret_yaml
+)
+
+
+__all__ = [
+    'SECRET_YAML', '_SECRET_NAMESPACE',
+    'dump', 'save_yaml',
+    'clear_secret_cache', 'load_yaml', 'secret_yaml',
+]
