+++ homeassistant/util/yaml/objects.py
@@ -1,0 +1,13 @@
+"""Custom yaml object types."""
+
+
+class NodeListClass(list):
+    """Wrapper class to be able to add attributes on a list."""
+
+    pass
+
+
+class NodeStrClass(str):
+    """Wrapper class to be able to add attributes on a string."""
+
+    pass
