+++ mypy.ini
@@ -17,7 +17,11 @@ disallow_untyped_defs = true
 [mypy-homeassistant.config_entries]
 disallow_untyped_defs = false
 
-[mypy-homeassistant.util.yaml]
+[mypy-homeassistant.util.yaml.dumper]
+warn_return_any = false
+disallow_untyped_calls = false
+
+[mypy-homeassistant.util.yaml.loader]
 warn_return_any = false
 disallow_untyped_calls = false
