+++ tests/components/scene/test_init.py
@@ -4,7 +4,7 @@ import unittest
 
 from homeassistant.setup import setup_component
 from homeassistant.components import light, scene
-from homeassistant.util import yaml
+from homeassistant.util.yaml import loader as yaml_loader
 
 from tests.common import get_test_home_assistant
 from tests.components.light import common as common_light
@@ -90,7 +90,7 @@ class TestScene(unittest.TestCase):
                 self.light_1.entity_id, self.light_2.entity_id)
 
         with io.StringIO(config) as file:
-            doc = yaml.yaml.safe_load(file)
+            doc = yaml_loader.yaml.load(file)
 
         assert setup_component(self.hass, scene.DOMAIN, doc)
         common.activate(self.hass, 'scene.test')
